Fix error in log_comet_images
